29/01/2025 1:58PM

+++ src/components/contest/commentForm.js
@@ -19,7 +19,7 @@ const CommentForm = ({ contestId, contest }) => {
     const [buttonComment, setButtonComment] = useState("Edit comment");
     const [likedComments, setLikedComments] = useState([]);
     const [statusUser, setStatusUser] = useState(false);
-    const [isFetchingLastestComments, , setIsFetchingLastestComments,] = useState(false);
+    const [isFetchingLastestComments, setIsFetchingLastestComments] = useState(false);
     const [isNewComment, setIsNewComment] = useState(false);
     const [editedComment, setEditedComment] = useState("");
 
@@ -224,7 +224,9 @@ const CommentForm = ({ contestId, contest }) => {
     };
 
     const handleDeleteComment = async () => {
-        alert("Are you sure you want to delete this comment?");
+        const confirmDelete = window.confirm("Are you sure you want to delete this comment?");
+        if (!confirmDelete) return;
+        
         try {
             await axios.delete(`http://localhost:5231/api/Contest/deleteComment`, {
                 params: { idComment: loggedAccountComment.idComment }
@@ -303,6 +305,7 @@ const CommentForm = ({ contestId, contest }) => {
                             value={loggedAccountComment.content}
                             onChange={(e) => setEditedComment(e.target.value)}
                         /> : loggedAccountComment.content}
+                        
                         <br />
                         <div style={{ display: "flex", flexDirection: "row", justifyContent: "space-between", marginTop: "10px" }}>
                             <div style={{ textAlign: "left" }}>
@@ -311,7 +314,8 @@ const CommentForm = ({ contestId, contest }) => {
                             </div>
                             <div style={{ textAlign: "right", display: "flex", flexDirection: "row", gap: "10px" }}>
                                 <span className="cmtForm-submit-button" onClick={handleDeleteComment}>Delete</span>
-                                <span className="cmtForm-submit-button" onClick={handleEditComment}>{buttonComment}</span>
+                                {!loggedAccountComment.isApproved &&
+                                    <span className="cmtForm-submit-button" onClick={handleEditComment}>{buttonComment}</span>}
                             </div>
                         </div>
                         <hr />
